Create parameter file for the 'JWSACruncher'

To run the 'JWSACruncher' needs a parameter file and create_param_file allows to create it.

Usage

create_param_file(
  dir_file_param,
  bundle = 10000,
  csv_layout = "list",
  csv_separator = ";",
  ndecs = 6,
  full_series_name = TRUE,
  rslt_name_level = 2,
  policy = "parameters",
  refreshall = TRUE,
  output = NULL,
  matrix_item = getOption("default_matrix_item"),
  tsmatrix_series = getOption("default_tsmatrix_series"),
  paths_path = NULL,
  v3 = getOption("is_cruncher_v3")
)

Arguments

dir_file_param

Path to the directory that will contains the parameter file "parameters.param".

bundle

Maximum size for a group of series (in output). By default bundle = 10000.

csv_layout

Layout of the CSV files (series only). By default csv_layout = "list". Other options: csv_layout = "vtable" (vertical table) or csv_layout = "htable" (horizontal table).

csv_separator

The field separator string used in the CSV file. By default csv_separator = ";".

ndecs

Number of decimals used in the output. By default ndec = 6.

full_series_name

Boolean indicating if the fully qualified name of the series will be used (the default full_series_name = TRUE) or if only the name of the series should be displayed. Only used when v3 = TRUE.

rslt_name_level

Only used when v3 = TRUE.

policy

Refreshing policy of the processing. By default policy = "parameters" (re-estimation of the coefficients of the reg-ARIMA model, see details).

refreshall

Boolean indicating if the data is refreshed (by default refreshall = TRUE).

output

Full path of the output folder. By default (output = NULL) a folder is create in the path to the workspace ([workspace]/Output).

matrix_item

Character containing the items of the matrix output (see the 'JDemetra+' manual for more information). By default, the items defined in the option getOption("default_matrix_item") are used (option initialized by the default output of the 'JWSACruncher' 2.2.2).

tsmatrix_series

Character containing the names of the times series to export (see the 'JDemetra+' manual for more information). By default, the items defined in the option getOption("default_tsmatrix_series") are used (option initialized by the default output of the 'JWSACruncher' 2.2.2).

paths_path

The paths used for relative addresses (see the "Demetra Paths" of the graphical interface of 'JDemetra+').

v3

Boolean indicating if the parameter file should be compatible with a version 3.0.0 and higher of 'JWSACRuncher' (v3 = TRUE) or a lower version (v3 = FALSE). By default the value of the option "is_cruncher_v3" is used (equals to FALSE by default).

Value

Path to the parameter file.

Details

When the 'JWSACruncher' is launched, the data is refreshed with a specific policy that is defined by the parameter policy. The available options are:

policy = "current": all the estimations are fixed;

policy = "fixedparameters" or policy = "fixed": re-estimation of the coefficients of the regression variables (but not the ARIMA coefficients);

policy = "parameters" (the default): policy = "fixedparameters" + re-estimation of ARIMA coefficients;

policy = "lastoutliers": policy = "parameters" + re-identification of last outliers (on the last year);

policy = "outliers": policy = "lastoutliers" + re-identification of all outliers;

policy = "stochastic": policy = "outliers" + re-identification of ARIMA orders;

policy = "complete" or policy = "concurrent": the model is completely re-identified and re-estimated.

See also

Examples

if (FALSE) {
dir = tempdir()
# Here a file parameters.param is created in the directory dir
# with default parameters of the different options
create_param_file(dir)
# to only export the raw and the seasonally adjusted series
 create_param_file(dir, 
                   tsmatrix_series = c("y", "sa"))
}